requirements.txt
loguru
mkdocs
pip
python-dotenv
ruff
tqdm
typer
selenium
beautifulsoup4
pandas
pypdf
streamlit

# LangChain & RAG Stack
langchain==0.3.27
langchain-openai==0.3.35
langgraph==0.6.11
langfuse==3.7.0
langchain-community

# SQL & Databases
sqlalchemy==2.0.45
psycopg2-binary==2.9.11
pymongo==4.15.5

# Vector Stores & RAG Components
chromadb==1.3.6
faiss-cpu==1.13.0
pinecone-client==6.0.0
weaviate-client==4.18.3
qdrant-client
langchain-qdrant

# Embeddings
sentence-transformers==5.1.2

# Chunking Pipelines
tiktoken
python-frontmatter
playwright